TodoListViewModelファイルを削除してTodoViewModelファイルに統合
