+++ QtlTodoList/Screen/Todo/TodoViewModel.swift
@@ -6,10 +6,61 @@
 //
 
 import Foundation
+import FirebaseFirestore
 
 class TodoViewModel: ObservableObject {
     // MARK: - Property Wrappers
     @Published var todo = Todos(title: "", message: "")
+    @Published var todos = [Todos]()
     // MARK: - Properties
     static let shared = TodoViewModel()
+    private let firestore = Firestore.firestore()
 }
+
+extension TodoViewModel {
+    // MARK: - Methods
+    /// Firestoreにデータ追加
+    func createFirestoreData(title: String, message: String) {
+        let todos = Todos(title: "", message: "")
+        Task {
+            try await firestore.collection("todos").document(todos.id).setData([
+                "title": title,
+                "message": message
+            ])
+        }
+    }
+
+    /// Firestoreのデータ読み込み
+    func readFirestoreData() {
+         firestore.collection("todos").addSnapshotListener { querySnapshot, error in
+            guard let documents = querySnapshot?.documents else {
+                print("error: \(error.debugDescription)")
+                return
+            }
+            self.todos = documents.map { querySnapshot -> Todos in
+                let data = querySnapshot.data()
+                let title = data["title"] as? String ?? ""
+                let message = data["message"] as? String ?? ""
+
+                return Todos(id: querySnapshot.documentID, title: title, message: message)
+            }
+         }
+    }
+
+    /// FireStoreのデータ更新
+    func updateFirestoreData(todo: Todos) {
+        Task {
+            try await firestore.collection("todos").document(todo.id).setData([
+                "title": todo.title,
+                "message": todo.message
+            ])
+        }
+    }
+
+    /// Firestoreのデータ削除
+    func deleteFirestoreData(todo: Todos) {
+        Task {
+            try await firestore.collection("todos").document(todo.id).delete()
+        }
+    }
+} // extension
